app refactoring: class, compose typeScript
