--- src/components/Nav/NavContainer.jsx
@@ -1,17 +1,0 @@
-import Nav from "./Nav";
-import { connect } from 'react-redux';
-// import { connect } from 'react-redux/es/exports';
-
-
-
-let mapStateToProps = (state) => {
-    return {
-        friendsData: state.sidebar.friendsData 
-    }
-}
-
-const NavContainer = connect(mapStateToProps)(Nav)
-
-
-
-export default NavContainer;
